Private value front-end, self-error uncertainty mode, fp32 value CE
From the 2026-08-09 comparative audit of value-training semantics vs
lczero-training (vdapda branch):

- PRIVATE VALUE FRONT-END (ValueHeadChannels, 0 = legacy): the value
  family (value1, value2, unc, HL-Gauss) gets its own per-square
  projection D -> C (flattened 64*C, reference C=32 => 2048-dim private
  input), bypassing the policy-shared doubly-compressed headPremap/
  headSharedLinear bottleneck (typ. 256-dim, optimized predominantly by
  the policy gradient). Audit finding #1 and the suspected structural
  seat of the persistent value deficit: every prior observation (value
  not scaling with model size, zero-effect value-only fine-tunes, the
  late-phase policy/value see-saw, loss-mass increases not helping) is
  consistent with the value head drinking through policy's straw.
  Served-graph change: resume across the toggle fails loudly (strict
  load), deliberately kept OUT of the aux-head reconciliation.
  Incompatible with vda/gtab modes and head-front LoRA (asserted).
- UNC SELF-ERROR mode (UncSelfError): unc trains toward the student's
  own realized (q_pred - q_target)^2, detached (lc0 ValueErrorLoss
  semantics) instead of the teacher-time |DeltaQVersusV| data field;
  optimistic-policy sigma switches to sqrt(unc). Serving note: exported
  'unc' output changes units (sigma^2) in this mode.
- fp32 cast of value CE logits/targets (both the standard and focal
  paths) - lc0 guards this; closes bf16-pure quantization at the
  late-run 0.01-nat KL scale.

Validated: 3M run with the full package (private C=32 + self-error +
soft/optimistic/HL-Gauss/replay/EMA) - clean boot, zero NaN, EXIT
SUCCESS. Reviewed; head-width migration verified complete across all
17 heads/consumers.

+++ src/CeresTrainPy/losses.py
@@ -251,6 +251,12 @@ class LossCalculator():
     """Cross entropy with optional per-record z-provenance weighting
     (CERES_VALUE_PROV_WEIGHTS). Falls back to plain mean CE when weighting is
     off or the batch carries no provenance (v7x-less shard in auto mode)."""
+    # fp32 cast on the logits (2026-08-09 comparative audit vs lc0, row 8):
+    # under BFloat16Pure the value logits would otherwise enter the CE already
+    # quantized to bf16 (~±0.004), non-negligible at the late-run 0.01-nat KL
+    # scale. lc0 guards this with an explicit fp32 cast in the value head.
+    output = output.float()
+    target = target.float()
     if VALUE_PROV_WEIGHTS is not None and provenance is not None:
       prov_wt = torch.tensor(VALUE_PROV_WEIGHTS, device=output.device, dtype=torch.float32)
       rec_wt = prov_wt[provenance.reshape(-1).long()]
@@ -270,7 +276,7 @@ class LossCalculator():
       # Hardness-weighted per-sample KL (see module header). Per-sample entropy
       # is subtracted BEFORE weighting so the hardness measure is the true KL,
       # not inflated by high-entropy (drawish) targets.
-      _ce_i = F.cross_entropy(output, target, reduction='none')
+      _ce_i = F.cross_entropy(output.float(), target.float(), reduction='none')
       _t = torch.clamp(target.float() + 1e-6, min=1e-6)
       _h_i = -(_t * torch.log(_t)).sum(dim=-1)
       _kl_i = _ce_i - _h_i
